[.Net] ollama agent improvements and reorganization
