+++ dotnet/src/Autogen.Ollama/Autogen.Ollama.csproj
@@ -9,8 +9,4 @@
       <ProjectReference Include="..\AutoGen.Core\AutoGen.Core.csproj" />
     </ItemGroup>
 
-    <ItemGroup>
-      <PackageReference Include="System.Net.Http.Json" Version="8.0.0" />
-    </ItemGroup>
-
 </Project>
