+++ dotnet/src/Autogen.Ollama/DTOs/ModelReplyOptions.cs
