+++ dotnet/src/Autogen.Ollama/DTOs/OllamaReplyOptions.cs
